review: [fra-sym] replace unit test for integration test
