+++ alma/phpunit.ci.xml
@@ -11,8 +11,8 @@
          stopOnFailure               = "true">
 
     <testsuites>
-        <testsuite name="Alma Installment Prestashop Unit Test Suite">
-            <directory>tests/unit</directory>
+        <testsuite name="Alma Installment Prestashop Integration Test Suite">
+            <directory>tests/integration</directory>
         </testsuite>
     </testsuites>
